requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
nltk
pandas==1.1.0
numpy
six